refactor: prepare separating htmlwidget
